fix(client-ng): make operation log its own module
This will allow easier testing.
